Clean up code and add new stuff
Add new environment where actions with no impact are masked - this env does not bind on dataset but STAYS on dataset AT ALL TIMES
Add action masking across the board in Networks, metrics and such so everything is compatible
Add precomputation of masks for environments - if no mask is found in files, precompute it on the fly. This is long however, but only happens once.